src/agent/requirements.txt
langchain==0.2.16
langchain-openai==0.1.23
langchain-community==0.2.16
openai==1.40.0
fastapi==0.112.0
uvicorn[standard]==0.30.6
google-cloud-storage==2.18.2
pydantic==2.8.2
httpx==0.27.0
